+++ crates/starknet-replay/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod block_context;
 pub mod rpc_state_reader;
+pub mod transactions;
 pub(crate) mod utils;
