+++ crates/starknet-replay/tests/test_replay_block.rs
@@ -1,117 +1,12 @@
-use std::error::Error;
-use std::sync::Arc;
-
 use blockifier::state::cached_state::CachedState;
-use blockifier::transaction::account_transaction::AccountTransaction;
 use blockifier::transaction::transactions::ExecutableTransaction;
 use rstest::rstest;
-use starknet::core::types::{BlockId, BlockWithTxs, Felt, InvokeTransaction, Transaction};
+use starknet::core::types::{BlockId, BlockWithTxs};
 use starknet::providers::jsonrpc::HttpTransport;
 use starknet::providers::{JsonRpcClient, Url};
-use starknet_api::core::PatriciaKey;
-use starknet_api::hash::StarkFelt;
-use starknet_api::stark_felt;
-use starknet_api::transaction::TransactionHash;
 use starknet_replay::block_context::build_block_context;
 use starknet_replay::rpc_state_reader::AsyncRpcStateReader;
-
-pub fn felt_vm2api(felt: Felt) -> StarkFelt {
-    stark_felt!(felt.to_hex_string().as_str())
-}
-
-pub(crate) fn starknet_rs_to_blockifier(
-    sn_core_tx: &starknet::core::types::Transaction,
-) -> Result<blockifier::transaction::transaction_execution::Transaction, Box<dyn Error>> {
-    // Map starknet_api transaction to blockifier's
-    let blockifier_tx: AccountTransaction = match sn_core_tx {
-        Transaction::Invoke(tx) => {
-            let (tx_hash, api_tx) = match tx {
-                InvokeTransaction::V0(tx) => {
-                    let _tx_hash = TransactionHash(felt_vm2api(tx.transaction_hash));
-                    unimplemented!();
-                }
-                InvokeTransaction::V1(tx) => {
-                    let tx_hash = TransactionHash(felt_vm2api(tx.transaction_hash));
-                    let api_tx = starknet_api::transaction::InvokeTransaction::V1(
-                        starknet_api::transaction::InvokeTransactionV1 {
-                            max_fee: starknet_api::transaction::Fee(tx.max_fee.to_biguint().try_into()?),
-                            signature: starknet_api::transaction::TransactionSignature(
-                                tx.signature.clone().into_iter().map(felt_vm2api).collect(),
-                            ),
-                            nonce: starknet_api::core::Nonce(felt_vm2api(tx.nonce)),
-                            sender_address: starknet_api::core::ContractAddress(
-                                PatriciaKey::try_from(felt_vm2api(tx.sender_address)).unwrap(),
-                            ),
-                            calldata: starknet_api::transaction::Calldata(Arc::new(
-                                tx.calldata.clone().into_iter().map(felt_vm2api).collect(),
-                            )),
-                        },
-                    );
-                    (tx_hash, api_tx)
-                }
-                InvokeTransaction::V3(tx) => {
-                    let _tx_hash = TransactionHash(felt_vm2api(tx.transaction_hash));
-                    unimplemented!();
-                }
-            };
-            let invoke =
-                blockifier::transaction::transactions::InvokeTransaction { tx: api_tx, tx_hash, only_query: false };
-            AccountTransaction::Invoke(invoke)
-        }
-        Transaction::DeployAccount(_tx) => {
-            // let contract_address = calculate_contract_address(
-            // tx.contract_address_salt(),
-            // tx.class_hash(),
-            // &tx.constructor_calldata(),
-            // ContractAddress::default(),
-            // )
-            // .unwrap();
-            // AccountTransaction::DeployAccount(DeployAccountTransaction {
-            // only_query: false,
-            // tx,
-            // tx_hash,
-            // contract_address,
-            // })
-            unimplemented!("starknet_rs_tx_to_blockifier() with Deploy txn");
-        }
-        Transaction::Declare(_tx) => {
-            /*
-            // Fetch the contract_class from the next block (as we don't have it in the previous one)
-            let next_block_state_reader = RpcStateReader(
-                RpcState::new_rpc(network, (block_number.next()).unwrap().into()).unwrap(),
-            );
-            let contract_class = next_block_state_reader
-                .get_compiled_contract_class(tx.class_hash())
-                .unwrap();
-            let class_info = calculate_class_info_for_testing(contract_class);
-            let declare = DeclareTransaction::new(tx, tx_hash, class_info).unwrap();
-            AccountTransaction::Declare(declare)
-            */
-            unimplemented!("starknet_rs_tx_to_blockifier() with Declare txn");
-        }
-        Transaction::L1Handler(_tx) => {
-            /*
-            // As L1Hanlder is not an account transaction we execute it here and return the result
-            let blockifier_tx = L1HandlerTransaction {
-                tx,
-                tx_hash,
-                paid_fee_on_l1: starknet_api::transaction::Fee(u128::MAX),
-            };
-            return (
-                blockifier_tx
-                    .execute(&mut state, &block_context, true, true)
-                    .unwrap(),
-                trace,
-                receipt,
-            );
-            */
-            unimplemented!("starknet_rs_tx_to_blockifier() with L1Handler txn");
-        }
-        _ => unimplemented!(),
-    };
-
-    Ok(blockifier::transaction::transaction_execution::Transaction::AccountTransaction(blockifier_tx))
-}
+use starknet_replay::transactions::starknet_rs_to_blockifier;
 
 #[rstest]
 #[ignore = "Requires a local Pathfinder node"]
